[improvement] Sped up fetching string-references (40m)
`Xlsx.index!` finds elements in an array by calling `strings.index`. Since the array is unsorted, Ruby has to do a slow sequence scan of the array to see if a string already exists.

Replacing the array with a hash makes checking to see if a string exists in the data structure _much_ faster.